drivers/iommu: Break out GSCID allocation
In preparation for operating multiple IOMMUs, break out the GSCID
allocation to be backed by a dedicated global allocation table. This
change just moves the existing code around, but there's probably an
opportunity here to switch to an alternative API that hands out
ref-counted RAII handles representing allocated GSCIDs.
